+++ mitra_services/src/ethereum/mod.rs
@@ -1,10 +1,2 @@
-mod api;
-pub mod contracts;
 pub mod eip4361;
-mod errors;
-pub mod signatures;
-pub mod sync;
 pub mod utils;
-
-pub use api::{EthereumApi, EthereumContract};
-pub use errors::EthereumError;