+++ scripts/build_deb.sh
@@ -37,10 +37,6 @@ mkdir -p $PACKAGE_DIR/usr/bin
 cp $TARGET_DIR/mitra $PACKAGE_DIR/usr/bin/mitra
 cp $TARGET_DIR/mitractl $PACKAGE_DIR/usr/bin/mitractl
 
-# Contracts
-mkdir -p $PACKAGE_DIR/usr/share/mitra
-cp -r contracts $PACKAGE_DIR/usr/share/mitra/contracts
-
 # Webapp
 mkdir -p $PACKAGE_DIR/usr/share/mitra
 # https://people.debian.org/~neilm/webapps-policy/ch-issues.html#s-issues-fhs