+++ src/main.rs
@@ -188,13 +188,6 @@ async fn main() -> std::io::Result<()> {
                 web::resource("/.well-known/{path}")
                     .to(HttpResponse::NotFound)
             );
-        if let Some(ethereum_config) = config.ethereum_config() {
-            // Serve artifacts if available
-            app = app.service(actix_files::Files::new(
-                "/contracts",
-                &ethereum_config.contract_dir,
-            ));
-        };
         if let Some(ref web_client_dir) = config.web_client_dir {
             app = app.service(web_client::static_service(web_client_dir));
         };