+++ CHANGELOG.md
@@ -14,6 +14,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Removed token gate.
 - Removed Ethereum subscriptions.
+- Disabled Etherem blockchain synchronization.
+- Removed `update-current-block` command.
+- Removed `generate-ethereum-address` command.
+- Removed `contract_address`, `features.gate` and `features.miner` from instance info.
 
 ## [2.26.0] - 2024-07-20